CTexture::Load 부분 문제

+++ Client/CScene_Start.cpp
@@ -40,7 +40,7 @@ void CScene_Start::Enter()
 	{
 		// Monster Object Ãß°¡
 		pMonsterObj = new CMonster;
-		pMonsterObj->SetPos(Vec2(blink + fTerm * i, 70.f + sinf(i) * 30));
+		pMonsterObj->SetPos(Vec2(blink + fTerm * i, 70.f + sin(i) * 30));
 		pMonsterObj->SetScale(Vec2(fObjScale, fObjScale));
 		pMonsterObj->SetCenterPos(pMonsterObj->GetPos());
 		pMonsterObj->SetMaxDistande(fMoveDist);
